tools/Sovereign/requirements.txt
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dnspython==2.7.0
idna==3.10
paramiko==3.5.0
prompt_toolkit==3.0.48
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.3
urllib3==2.2.3
wcwidth==0.2.13
tqdm==4.66.4
aiohttp==3.10.2
